tests/compliance/nist_pqc_test.py
# NIST PQC 合规性测试示例
import pytest

NIST_KYBER_VECTORS = [b"vector1", b"vector2"]

def validate_kyber(vector):
    # 伪实现，实际应调用 PQC 验证库
    return True

def test_kyber_nist_vectors():
    for vector in NIST_KYBER_VECTORS:
        assert validate_kyber(vector)