pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qacmf"
version = "2.0.0"
description = "Quantum-Aware Cryptography Migration Framework"
authors = [
    {name = "QACMF Team", email = "opensource@qacmf.org"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "cryptography>=41.0.0",
    "pyyaml",
    "requests"
]

[project.urls]
"Homepage" = "https://github.com/crystal-tensor/QACMF"
"Documentation" = "https://github.com/crystal-tensor/QACMF/tree/main/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"qacmf" = ["*.py", "*.yaml", "*.json"]